+++ test/easy_localization_widget_test.dart
@@ -77,6 +77,25 @@ void main() {
   );
 
   testWidgets(
+    '[EasyLocalization with  child==null  AssertionError] test',
+    (WidgetTester tester) async {
+      await tester.runAsync(() async {
+        try {
+          await tester.pumpWidget(
+          EasyLocalization(
+          child: null,
+          path: 'i18n',
+          supportedLocales: [Locale('en', 'US')],
+        ));
+        } on AssertionError catch (e) {
+          // throw  AssertionError('Expected ArgumentError');
+          expect(e, isAssertionError);
+        }
+      });
+    },
+  );
+
+  testWidgets(
     '[EasyLocalization with  RootBundleAssetLoader] test',
     (WidgetTester tester) async {
       await tester.runAsync(() async {
@@ -362,7 +381,7 @@ void main() {
           saveLocale: false,
           useOnlyLangCode: true,
           supportedLocales: [Locale('ar')],
-          fallbackLocale:Locale('ar') ,
+          fallbackLocale: Locale('ar'),
         ));
         await tester.idle();
         // The async delegator load will require build on the next frame. Thus, pump
@@ -408,6 +427,28 @@ void main() {
     });
 
     testWidgets(
+      '[EasyLocalization] SavedLocale()  null locale without country code',
+      (WidgetTester tester) async {
+        await tester.runAsync(() async {
+          await tester.pumpWidget(EasyLocalization(
+            child: MyApp(),
+            path: 'i18n',
+            // fallbackLocale:Locale('en') ,
+            supportedLocales: [Locale('en'), Locale('ar')], //
+          ));
+          await tester.idle();
+          await tester.pump(Duration(seconds: 2));
+          // The async delegator load will require build on the next frame. Thus, pump
+          await tester.pumpAndSettle();
+
+          expect(EasyLocalization.of(_context).supportedLocales,
+              [Locale('en'), Locale('ar')]);
+          expect(EasyLocalization.of(_context).locale, Locale('en'));
+          expect(EasyLocalization.of(_context).fallbackLocale, null);
+        });
+      },
+    );
+    testWidgets(
       '[EasyLocalization] SavedLocale()  test',
       (WidgetTester tester) async {
         await tester.runAsync(() async {
@@ -428,7 +469,8 @@ void main() {
           expect(EasyLocalization.of(_context).fallbackLocale, null);
         });
       },
-    );testWidgets(
+    );
+    testWidgets(
       '[EasyLocalization] startLocale  test',
       (WidgetTester tester) async {
         await tester.runAsync(() async {